feat, docs, debug: Implementation of the 'buttonAction' function and little modification of some vars in the calculatorFuncionality.js script.
feat --> Implementation of the 'buttonAction' function (still
incomplete, but started). This function calls another one that it
doesn't exist yet in the JS script.
Also implementation of more text in the 'notice(privacy)' function in the JS script.

docs --> Modification of several functions documentation in the JS
script.

debug --> Little improvement in the definition of the 'colors' array.
Now it's a CONST and not a VAR.